src/requirements.txt
streamlit==1.29.0
mysql-connector-python==8.2.0
pandas==2.1.4
plotly==5.18.0
validate-docbr==1.10.0
python-dotenv==1.0.0
